backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
openai==1.35.0
langchain==0.2.5
langchain-community==0.2.5
faiss-cpu==1.8.0
pypdf==4.2.0
sentence-transformers==3.0.1
numpy==1.26.4
pydantic==2.7.4
